Update swagger run script

+++ swagger/run-swagger-docs.sh
@@ -2,6 +2,7 @@
 # https://github.com/swagger-api/swagger-ui/blob/master/docs/usage/installation.md
 # https://hub.docker.com/r/swaggerapi/swagger-ui/tags
 
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 API_VER=$(curl -s https://api.github.com/repos/OasisLMF/OasisPlatform/tags | jq -r '( first ) | .name')
 API_URL="https://github.com/OasisLMF/OasisPlatform/releases/download/$API_VER/openapi-schema-$API_VER.json"
 
@@ -10,7 +11,11 @@ export SWAGGER_LINK="oasis-api-schema.json"
 export SWAGGER_IMAGE="swaggerapi/swagger-ui"
 export SWAGGER_TAG="v3.24.2"
 
-wget $API_URL -O $SWAGGER_SCHEMA
-rm $SWAGGER_LINK
-ln -s $SWAGGER_SCHEMA $SWAGGER_LINK
+cd $SCRIPT_DIR
+if [ ! -f $SWAGGER_SCHEMA ]; then
+  wget $API_URL -O $SWAGGER_SCHEMA
+  rm $SWAGGER_LINK
+  ln -s $SWAGGER_SCHEMA $SWAGGER_LINK
+fi
+
 docker-compose up -d
